+++ CHANGELOG.md
@@ -122,6 +122,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Diff
   - `current`: didn't apply global variants
   - `current`: didn't honor KiCad native DNP flags, they need a filter
+- PCB Print:
+  - Issues when forcing the board edge and using scaling (#532)
 
 
 ## [1.6.3] - 2023-06-26